Run RPC completeness tests where their toolchain is provisioned
The C# and Go completeness tests parse via the out-of-process tool/binary, which
is only built for the integTest source set (goBuild / csharpBuild); in the plain
test task the Go binary is absent and the C# path resolves an unpublished tool
version from NuGet. Move both to integTest. Also parse all JavaScript snippets
through a single parser/parseInputs call rather than building 27 parsers, which
was crashing the RPC process in CI.
